refactor: cria hierarquia de herança com PessoaBase para entidades de pessoa
- Adiciona classe abstrata PessoaBase herdando de BaseEntity
- Move BaseEntity para namespace DevXpert.Store.Core.Business.Models.Base
- Refatora Cliente e Vendedor para herdar de PessoaBase
- Remove propriedades duplicadas (Nome, Email, Senha) das entidades
- Atualiza imports em todas as classes afetadas
- Melhora organização da arquitetura seguindo princípio DRY

BREAKING CHANGE: BaseEntity movido para namespace Base
