+++ src/Backend/DevXpert.Store.Core/Business/Models/Base/PessoaBase.cs
@@ -1,0 +1,8 @@
+namespace DevXpert.Store.Core.Business.Models.Base;
+
+public class PessoaBase : BaseEntity
+{
+    public string Nome { get; set; }
+    public string Email { get; set; }
+    public string Senha { get; set; }
+}
