+++ src/Backend/DevXpert.Store.Core/Business/Models/Categoria.cs
@@ -1,4 +1,5 @@
-﻿using DevXpert.Store.Core.Business.Validations;
+﻿using DevXpert.Store.Core.Business.Models.Base;
+using DevXpert.Store.Core.Business.Validations;
 
 namespace DevXpert.Store.Core.Business.Models
 {
