+++ src/Backend/DevXpert.Store.Core/Business/Models/Cliente.cs
@@ -1,13 +1,10 @@
-﻿using DevXpert.Store.Core.Business.Validations;
+﻿using DevXpert.Store.Core.Business.Models.Base;
+using DevXpert.Store.Core.Business.Validations;
 
 namespace DevXpert.Store.Core.Business.Models
 {
-    public class Cliente : BaseEntity
+    public class Cliente : PessoaBase
     {
-        public string Nome { get; set; }
-        public string Email { get; set; }
-        public string Senha { get; set; }
-
         #region NAVIGATION PROPERTIES
        //TODO: MAPEAR LISTA DE FAVORITOS
         #endregion
